Handle "type var = ...", ++, and --
This change adds handling for variable declarations that contain an
assignment, which won't register as BinaryOperators, but instead will
be a VarDecl with an initializer. As well, capture both the ++ and --
as i & d, respectively - to record increment/decrement operations. In
this version, handle prefix & postfix identically, which should
introduce some fuzz.