Add a Ghidra plugin to do "PCode op List" -> "signature string"
This is intended to be a PCode analogue to the StmtParser.cpp Clang
front-end plugin. This Ghidra plugin will collect up the sequence of
PCode operations, and write them out to a JSON object where each key
is the function name, and each value is a list of the opcodes from
beginning to end. Optionally, a feature also exists to output a
condensed form of this that matches the function signature string output
from StmtParser.cpp.